Revise TOML format
Instead of an `[Inventory]` section, the TOML format now has to start
with a comment line that declares the format version. The header fields
`project` and `version` are now top-level keys.

This looks cleaner, saves a few bytes, and introduces unique "magic
bytes" at the beginning of the file so that an `inventory.toml` file
could be recognized as a distinct fromat by FileIO.jl.

Also, `version` is now optional.

Most previous (experimental) `inventory.toml` files can still be read
without error.